+++ helpers.cpp
@@ -1,0 +1,20 @@
+namespace helpers
+{
+
+int sum_of_digits(int num){
+    int result = 0;
+    do {
+        result = result + num%10;
+        num=num/10;
+    }
+    while(num!=0);
+    return result;
+}
+
+bool is_more_than(int number, int minimum)
+{
+    return (number > minimum);
+}
+
+} // namespace helpers
+
